requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SocketIO==5.3.5
Flask-SQLAlchemy==3.1.1
python-socketio==5.10.0
eventlet==0.33.3
requests==2.31.0
python-dotenv==1.0.0
PyJWT==2.8.0
werkzeug==3.0.1
